requirements.txt
blinker==1.8.2
branca==0.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-WTF==1.2.1
folium==0.17.0
fonttools==4.53.0
geopandas==1.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
mapclassify==2.7.0
MarkupSafe==2.1.5
matplotlib==3.9.0
networkx==3.3
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.4
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
xyzservices==2024.6.0
